Improved inlusion layout, and improved library compilation.

+++ flitdb/flit.c
@@ -29,7 +29,7 @@ flitdb_extern char *flitdb_errmsg(flitdb **handler)
 	return flitdb_get_err_message(handler);
 }
 
-flitdb_extern int flitdb_extract(flitdb **handler, uint64_t column_position, uint64_t row_position)
+flitdb_extern int flitdb_extract(flitdb **handler, unsigned short column_position, unsigned short row_position)
 {
 	return flitdb_read_at(handler, column_position, row_position);
 }
@@ -39,7 +39,7 @@ int flitdb_retrieved_type(flitdb **handler)
 	return flitdb_retrieve_value_type(handler);
 }
 
-flitdb_extern int flitdb_insert_int(flitdb **handler, uint64_t column_position, uint64_t row_position, signed long long int value)
+flitdb_extern int flitdb_insert_int(flitdb **handler, unsigned short column_position, unsigned short row_position, signed long long int value)
 {
 	flitdb_insert_value_int(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -47,14 +47,14 @@ flitdb_extern int flitdb_insert_int(flitdb **handler, uint64_t column_position, 
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, uint64_t column_position, uint64_t row_position, signed long long int value)
+int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, signed long long int value)
 {
 	return flitdb_insert_int(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_double(flitdb **handler, uint64_t column_position, uint64_t row_position, long double value)
+flitdb_extern int flitdb_insert_double(flitdb **handler, unsigned short column_position, unsigned short row_position, long double value)
 {
 	flitdb_insert_value_double(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -62,14 +62,14 @@ flitdb_extern int flitdb_insert_double(flitdb **handler, uint64_t column_positio
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, uint64_t column_position, uint64_t row_position, long double value)
+int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, long double value)
 {
 	return flitdb_insert_double(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_float(flitdb **handler, uint64_t column_position, uint64_t row_position, float value)
+flitdb_extern int flitdb_insert_float(flitdb **handler, unsigned short column_position, unsigned short row_position, float value)
 {
 	flitdb_insert_value_float(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -77,14 +77,14 @@ flitdb_extern int flitdb_insert_float(flitdb **handler, uint64_t column_position
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, uint64_t column_position, uint64_t row_position, float value)
+int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, float value)
 {
 	return flitdb_insert_float(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_char(flitdb **handler, uint64_t column_position, uint64_t row_position, char *value)
+flitdb_extern int flitdb_insert_char(flitdb **handler, unsigned short column_position, unsigned short row_position, char *value)
 {
 	flitdb_insert_value_char(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -92,14 +92,14 @@ flitdb_extern int flitdb_insert_char(flitdb **handler, uint64_t column_position,
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, uint64_t column_position, uint64_t row_position, char *value)
+int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, char *value)
 {
 	return flitdb_insert_char(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_const_char(flitdb **handler, uint64_t column_position, uint64_t row_position, const char *value)
+flitdb_extern int flitdb_insert_const_char(flitdb **handler, unsigned short column_position, unsigned short row_position, const char *value)
 {
 	int value_length = strlen(value);
 	char tmp_value[(value_length + 1)];
@@ -111,14 +111,14 @@ flitdb_extern int flitdb_insert_const_char(flitdb **handler, uint64_t column_pos
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, uint64_t column_position, uint64_t row_position, const char *value)
+int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, const char *value)
 {
 	return flitdb_insert_const_char(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_insert_bool(flitdb **handler, uint64_t column_position, uint64_t row_position, bool value)
+flitdb_extern int flitdb_insert_bool(flitdb **handler, unsigned short column_position, unsigned short row_position, bool value)
 {
 	flitdb_insert_value_bool(handler, value);
 	return flitdb_insert_at(handler, column_position, row_position);
@@ -126,14 +126,14 @@ flitdb_extern int flitdb_insert_bool(flitdb **handler, uint64_t column_position,
 
 #ifdef __cplusplus
 
-int flitdb_insert(flitdb **handler, uint64_t column_position, uint64_t row_position, bool value)
+int flitdb_insert(flitdb **handler, unsigned short column_position, unsigned short row_position, bool value)
 {
 	return flitdb_insert_bool(handler, column_position, row_position, value);
 }
 
 #endif
 
-flitdb_extern int flitdb_delete(flitdb **handler, uint64_t column_position, uint64_t row_position)
+flitdb_extern int flitdb_delete(flitdb **handler, unsigned short column_position, unsigned short row_position)
 {
 	flitdb_insert_reset(handler);
 	return flitdb_insert_at(handler, column_position, row_position);
